# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gallery
    description: Gallery image operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gallery:
    get:
      operationId: getGalleryImages
      tags: [gallery]
      summary: List all gallery images
      description: Returns all gallery images ordered by creation date
      responses:
        "200":
          description: List of gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add a gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryImageRequest"
      responses:
        "201":
          description: Created gallery image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted gallery image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - imageUrl
        - title
        - description
        - createdAt
        - updatedAt
    CreateGalleryImageRequest:
      type: object
      properties:
        imageUrl:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - imageUrl
        - title
        - description
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error