# lib/api-zod/src/generated/types/galleryImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImage {
  id: number;
  imageUrl: string;
  title: string;
  description: string;
  createdAt: string;
  updatedAt: string;
}